+++ src/usable/effectsMakers.ts
@@ -1,25 +1,22 @@
 import { toSafeEffectId } from "../helpers/effects";
-import { AllState, ItemType, StepName } from "../types";
+import { AllState, ItemType, PropName, PropValue, StepName } from "../types";
 import { makeEffects } from "./effects";
 import { getPrevState, getState } from "./getSet";
 
 export function makeEffectsMaker<
-  T_StoreName extends ItemType & string,
-  T_StoreItemId extends keyof AllState[T_StoreName] & string,
-  T_PropertyName extends keyof AllState[T_StoreName][T_StoreItemId] & string,
+  T_ItemType extends ItemType & string,
+  T_PropName extends PropName<T_ItemType> & string,
+  T_PropValue extends EnsureString<PropValue<T_ItemType, T_PropName>>,
   T_StepName extends StepName,
   T_UsefulParams extends Record<any, any>
 >(
-  storeName: T_StoreName,
-  storeItemId: T_StoreItemId,
-  storyProperty: T_PropertyName,
+  storeName: T_ItemType,
+  storeItemId: string,
+  storyProperty: T_PropName,
   stepName?: T_StepName,
   getUsefulParams?: () => T_UsefulParams
 ) {
-  type StoreState = AllState[T_StoreName][T_StoreItemId];
-
-  type PropertyValue = StoreState[T_PropertyName];
-  type CallbacksMap = Partial<Record<PropertyValue, (usefulStuff: T_UsefulParams) => void>>;
+  type CallbacksMap = Partial<Record<EnsureString<T_PropValue>, (usefulStuff: T_UsefulParams) => void>>;
   const mainEffectId = toSafeEffectId(`customEffectFor_${storeName}_${storyProperty}`);
 
   function newEffectMaker(callbacksMap: CallbacksMap) {
@@ -27,7 +24,7 @@ export function makeEffectsMaker<
       whenPropertyChanges: makeEffect(
         (_, _diffInfo) => {
           const usefulParams = getUsefulParams?.();
-          const latestValue = getState(storeName, storeItemId)[storyProperty] as PropertyValue;
+          const latestValue = getState(storeName, storeItemId)[storyProperty] as T_PropValue;
 
           callbacksMap[latestValue]?.(usefulParams!);
         },
@@ -47,29 +44,26 @@ export function makeEffectsMaker<
 }
 
 export function makeLeaveEffectsMaker<
-  T_StoreName extends ItemType & string,
-  T_StoreItemId extends keyof AllState[T_StoreName] & string,
-  T_PropertyName extends keyof AllState[T_StoreName][T_StoreItemId] & string,
+  T_ItemType extends ItemType & string,
+  T_PropName extends PropName<T_ItemType>,
+  T_PropValue extends EnsureString<PropValue<T_ItemType, T_PropName>>,
   T_StepName extends StepName,
   T_UsefulParams extends Record<any, any>
 >(
-  storeName: T_StoreName,
-  storeItemId: T_StoreItemId,
-  storyProperty: T_PropertyName,
+  storeName: T_ItemType,
+  storeItemId: string,
+  storyProperty: T_PropName,
   stepName?: T_StepName,
   getUsefulParams?: () => T_UsefulParams
 ) {
-  type StoreState = AllState[T_StoreName][T_StoreItemId];
-
-  type PropertyValue = StoreState[T_PropertyName];
-  type RulesOptions = Partial<Record<PropertyValue, (usefulStuff: T_UsefulParams) => void>>;
+  type RulesOptions = Partial<Record<EnsureString<T_PropValue>, (usefulStuff: T_UsefulParams) => void>>;
   const mainEffectId = toSafeEffectId(`customEffectFor_${storeName}_${storyProperty}`);
   function newEffectsMaker(callBacksObject: RulesOptions) {
     return makeEffects((makeEffect) => ({
       whenPropertyChanges: makeEffect(
         (_, _diffInfo) => {
           const usefulStoryStuff = getUsefulParams?.();
-          const prevValue = getPrevState(storeName, storeItemId)[storyProperty] as PropertyValue;
+          const prevValue = getPrevState(storeName, storeItemId)[storyProperty] as T_PropValue;
 
           callBacksObject[prevValue]?.(usefulStoryStuff!);
         },
@@ -87,35 +81,33 @@ export function makeLeaveEffectsMaker<
 
   return newEffectsMaker;
 }
+type EnsureString<T> = T extends string ? T : never;
 
 /** Similar to makeEffectsMaker but accepts parameters for two store properties (can be from different stores) , and the callback fires when properties of both stores change */
 export function makeNestedEffectsMaker<
-  T_StoreName1 extends ItemType & string,
-  T_StoreItemId1 extends keyof AllState[T_StoreName1] & string,
-  T_PropName1 extends keyof AllState[T_StoreName1][T_StoreItemId1] & string,
-  T_StoreName2 extends ItemType & string,
-  T_StoreItemId2 extends keyof AllState[T_StoreName2] & string,
-  T_PropName2 extends keyof AllState[T_StoreName2][T_StoreItemId2] & string,
+  T_ItemType1 extends ItemType & string,
+  T_PropName1 extends PropName<T_ItemType1> & string,
+  T_PropValue1 extends EnsureString<PropValue<T_ItemType1, T_PropName1>>,
+  T_ItemType2 extends ItemType & string,
+  T_PropName2 extends PropName<T_ItemType2> & string,
+  T_PropValue2 extends EnsureString<PropValue<T_ItemType2, T_PropName2>>,
   T_StepName extends StepName,
   T_UsefulParams extends Record<any, any>
 >(
-  storeInfo1: [T_StoreName1, T_StoreItemId1, T_PropName1],
-  storeInfo2: [T_StoreName2, T_StoreItemId2, T_PropName2],
+  storeInfo1: [T_ItemType1, string, T_PropName1],
+  storeInfo2: [T_ItemType2, string, T_PropName2],
   stepName?: T_StepName,
   getUsefulParams?: () => T_UsefulParams
 ) {
   const [storeName1, storeItemId1, storyProp1] = storeInfo1;
   const [storeName2, storeItemId2, storyProp2] = storeInfo2;
 
-  type StoreState1 = AllState[T_StoreName1][T_StoreItemId1];
-  type StoreState2 = AllState[T_StoreName2][T_StoreItemId2];
-
-  type PropValue1 = StoreState1[T_PropName1];
-  type PropValue2 = StoreState2[T_PropName2];
   type RulesOptions = Partial<
     Record<
-      PropValue1,
-      Partial<Record<PropValue2, (usefulStuff: ReturnType<NonNullable<typeof getUsefulParams>>) => void>>
+      EnsureString<T_PropValue1>,
+      Partial<
+        Record<EnsureString<T_PropValue2>, (usefulStuff: ReturnType<NonNullable<typeof getUsefulParams>>) => void>
+      >
     >
   >;
   const mainEffectId = toSafeEffectId(`customEffectFor_${storeName1}_${storyProp1}_${storeName2}_${storyProp2}`);
@@ -124,8 +116,8 @@ export function makeNestedEffectsMaker<
       whenPropertyChanges: makeEffect(
         (_, diffInfo) => {
           const usefulStoryStuff = getUsefulParams?.();
-          const latestValue1 = getState(storeName1, storeItemId1)?.[storyProp1] as PropValue1;
-          const latestValue2 = getState(storeName2, storeItemId2)?.[storyProp2] as PropValue2;
+          const latestValue1 = getState(storeName1, storeItemId1)?.[storyProp1] as T_PropValue1;
+          const latestValue2 = getState(storeName2, storeItemId2)?.[storyProp2] as T_PropValue2;
 
           // Make sure value1 changed for itemType1 etc, since it listens for both itemIIds for both stores (might not be neccesary)
           const didValue1Change = diffInfo.propsChangedBool[storeName1][storeItemId1][storyProp1];
@@ -150,30 +142,23 @@ export function makeNestedEffectsMaker<
 
 /** The same as makeNestedRuleMaker , but the callback fires when the properties of both stores become NOT the specified values, but were previously */
 export function makeNestedLeaveEffectsMaker<
-  T_StoreName1 extends ItemType & string,
-  T_StoreItemId1 extends keyof AllState[T_StoreName1] & string,
-  T_PropertyName1 extends keyof AllState[T_StoreName1][T_StoreItemId1] & string,
-  T_StoreName2 extends ItemType & string,
-  T_StoreItemId2 extends keyof AllState[T_StoreName2] & string,
-  T_PropertyName2 extends keyof AllState[T_StoreName2][T_StoreItemId2] & string,
+  T_ItemType1 extends ItemType & string,
+  T_PropName1 extends PropName<T_ItemType1> & string,
+  T_ItemType2 extends ItemType & string,
+  T_PropName2 extends PropName<T_ItemType2> & string,
   T_UsefulParams extends Record<any, any>
 >(
-  storeInfo1: [T_StoreName1, T_StoreItemId1, T_PropertyName1],
-  storeInfo2: [T_StoreName2, T_StoreItemId2, T_PropertyName2],
+  storeInfo1: [T_ItemType1, string, T_PropName1],
+  storeInfo2: [T_ItemType2, string, T_PropName2],
   stepName?: StepName,
   getUsefulParams?: () => T_UsefulParams
 ) {
   const [storeName1, storeItemId1, storyProperty1] = storeInfo1;
   const [storeName2, storeItemId2, storyProperty2] = storeInfo2;
 
-  type StoreState1 = AllState[T_StoreName1][T_StoreItemId1];
-  type StoreState2 = AllState[T_StoreName2][T_StoreItemId2];
-
-  type PropertyValue1 = StoreState1[T_PropertyName1];
-  type PropertyValue2 = StoreState2[T_PropertyName2];
-  type RulesOptions = Partial<
-    Record<PropertyValue1, Partial<Record<PropertyValue2, (usefulStuff: T_UsefulParams) => void>>>
-  >;
+  type PropValue1 = PropValue<T_ItemType1, T_PropName1>;
+  type PropValue2 = PropValue<T_ItemType2, T_PropName2>;
+  type RulesOptions = Partial<Record<PropValue1, Partial<Record<PropValue2, (usefulStuff: T_UsefulParams) => void>>>>;
   const mainEffectId = toSafeEffectId(
     `customLeaveEffectFor_${storeName1}_${storyProperty1}_${storeName2}_${storyProperty2}`
   );
@@ -182,10 +167,10 @@ export function makeNestedLeaveEffectsMaker<
       whenPropertyChanges: makeEffect(
         (_, diffInfo) => {
           const usefulParams = getUsefulParams?.();
-          const latestValue1 = getState(storeName1, storeItemId1)?.[storyProperty1] as PropertyValue1;
-          const latestValue2 = getState(storeName2, storeItemId2)?.[storyProperty2] as PropertyValue2;
-          const prevValue1 = getPrevState(storeName1, storeItemId1)[storyProperty1] as PropertyValue1;
-          const prevValue2 = getPrevState(storeName2, storeItemId2)[storyProperty2] as PropertyValue2;
+          const latestValue1 = getState(storeName1, storeItemId1)?.[storyProperty1] as PropValue1;
+          const latestValue2 = getState(storeName2, storeItemId2)?.[storyProperty2] as PropValue2;
+          const prevValue1 = getPrevState(storeName1, storeItemId1)[storyProperty1] as PropValue1;
+          const prevValue2 = getPrevState(storeName2, storeItemId2)[storyProperty2] as PropValue2;
 
           // Make sure value1 changed for itemType1 etc, since it listens for both itemIIds for both stores (might not be neccesary)
           const didValue1Change = diffInfo.propsChangedBool[storeName1][storeItemId1][storyProperty1];