+++ CHANGELOG.md
@@ -8,6 +8,7 @@ TODO
 - look into addItem doing a setState if the item was already added
 - clean up the code a bit
 - maybe support starting param effects with group.name instead of seperate params
+- maybe remove "getUsefulParams" from prendy effect makers
 
 v1.0.0